Converted code to support system.text.json
Updated the code to support system.text.json, and also updated the tests
to run using system.text.json.

Issues that needs fixing:
Could not make the jsonconstructor work properly, so i updated the
properties to be INIT.

This does not solve that the classes can be instantiated invalidly, so
more work might have to go into solving the jsonconstructors.